requirements.txt
cleverbot
xkcd
discord.py[voice]
youtube-dl
aioredis
redis
pyfiglet
wolframalpha
praw
